fix: moved ErrorMessage interface to its componenet
