--- src/interfaces/errorMessageProps.ts
@@ -1,3 +1,0 @@
-export default interface ErrorMessageProps  {
-  message: string;
-}
