+++ src/interfaces/index.ts
@@ -1,3 +1,2 @@
-export type { default as ErrorMessageProps } from "./errorMessageProps";
-export type { default as Pokemon} from "./pokemonProps";
+export type { default as Pokemon} from "./pokemon";
 export type { default as Type } from "./type";
