+++ src/interfaces/pokemon.ts
